src/api/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
install==1.3.5
numpy==1.26.1
packaging==23.2
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
